formatted with rustfmt

+++ src/timetable.rs
@@ -1,22 +1,21 @@
 use crate::auth::login;
+use crate::class;
 use crate::models::Timetable;
 use crate::user;
-use crate::class;
+use chrono::NaiveDate;
+use dotenv::dotenv;
+use hyper::{Body, Client, Request};
+use hyper_tls::HttpsConnector;
+use serde_json;
+use serde_json::value::Value::Null;
 use sqlx::*;
 use std::env;
-use dotenv::dotenv;
-use serde_json;
-use hyper::{Client, Request, Body};
-use hyper_tls::HttpsConnector;
-use chrono::NaiveDate;
 use std::time::Instant;
-use serde_json::value::Value::Null;
 
 type Result<T> = std::result::Result<T, Box<dyn std::error::Error + Send + Sync>>;
 
 // TODO: Pass in a master PgPool for each DB function instead of instantiating a new one each time (takes a lot of time)
 pub async fn initialise_timetable(user_id: i64, pool: &PgPool) -> Result<String> {
-
     let synergetic_id = user::get_user_synergetic_id_by_user_id(user_id, pool).await?;
     let fetched_date = chrono::offset::Local::now().naive_local().date();
 
@@ -24,39 +23,58 @@ pub async fn initialise_timetable(user_id: i64, pool: &PgPool) -> Result<String>
 
     if timetable_id == 0 {
         timetable_id = create_timetable(user_id, fetched_date, pool).await?; // Create a new timetable
-        let fetched = fetch_timetable_by_synergetic_id(synergetic_id, user_id, timetable_id, pool).await?;
+        let fetched =
+            fetch_timetable_by_synergetic_id(synergetic_id, user_id, timetable_id, pool).await?;
         let fetched = fetched.as_str();
         match fetched {
             "successful" => println!("Timetable exists on MGS API - id:{}", synergetic_id),
-            _ => { return Ok("unsuccessful".to_string()); }
+            _ => {
+                return Ok("unsuccessful".to_string());
+            }
         }
 
         return Ok("successful".to_string());
     }
 
     class::delete_all_classes_in_timetable(timetable_id, pool).await?; // Delete existing classes
-    let fetched = fetch_timetable_by_synergetic_id(synergetic_id, user_id, timetable_id, pool).await?; // If timetable exists, get new set of classes and update fetched date
+    let fetched =
+        fetch_timetable_by_synergetic_id(synergetic_id, user_id, timetable_id, pool).await?; // If timetable exists, get new set of classes and update fetched date
     let fetched = fetched.as_str();
 
     // Check if timetable exists on myMGS API
     match fetched {
         "successful" => println!("Timetable exists on MGS API - id:{}", synergetic_id),
-        _ => { return Ok("unsuccessful".to_string()); }
+        _ => {
+            return Ok("unsuccessful".to_string());
+        }
     }
 
     update_timetable_by_user_id(user_id, fetched_date, pool).await?; // Update fetched date of existing timetable
     Ok("successful".to_string())
 }
 
-pub async fn fetch_timetable_by_synergetic_id(synergetic_id: i32, user_id: i64, timetable_id: i32, pool: &Pool<Postgres>) -> Result<String> {
+pub async fn fetch_timetable_by_synergetic_id(
+    synergetic_id: i32,
+    user_id: i64,
+    timetable_id: i32,
+    pool: &Pool<Postgres>,
+) -> Result<String> {
     dotenv().ok();
 
     let username = env::var("MGS_USERNAME").expect("USERNAME must be set"); // Using master credentials
     let password = env::var("MGS_PASSWORD").expect("PASSWORD must be set");
 
-    let timetable_api_url = format!("https://my.mgs.vic.edu.au/mg/api/timetable_week/{}/{}/S/student.json", synergetic_id, chrono::offset::Local::now().format("%Y-%m-%d"));
-    let (simple_saml_session_id, simple_saml_auth_token_cookie, ssess_cookie) = login(username.as_str(), password.as_str()).await?;
-    let cookie = format!("{}; {}; {}", simple_saml_session_id, simple_saml_auth_token_cookie, ssess_cookie);
+    let timetable_api_url = format!(
+        "https://my.mgs.vic.edu.au/mg/api/timetable_week/{}/{}/S/student.json",
+        synergetic_id,
+        chrono::offset::Local::now().format("%Y-%m-%d")
+    );
+    let (simple_saml_session_id, simple_saml_auth_token_cookie, ssess_cookie) =
+        login(username.as_str(), password.as_str()).await?;
+    let cookie = format!(
+        "{}; {}; {}",
+        simple_saml_session_id, simple_saml_auth_token_cookie, ssess_cookie
+    );
 
     let https = HttpsConnector::new();
     let client = Client::builder().build::<_, hyper::Body>(https);
@@ -72,7 +90,8 @@ pub async fn fetch_timetable_by_synergetic_id(synergetic_id: i32, user_id: i64, 
     println!("Got back timetable JSON: {}ms", now.elapsed().as_millis());
     let body_bytes = hyper::body::to_bytes(response).await?;
     let body = String::from_utf8(body_bytes.to_vec()).expect("response was not valid utf-8");
-    let json: serde_json::Value = serde_json::from_str(&body.as_str()).expect("JSON was not formatted properly");
+    let json: serde_json::Value =
+        serde_json::from_str(&body.as_str()).expect("JSON was not formatted properly");
     let mut d_number = 1;
 
     if json == Null {
@@ -91,7 +110,8 @@ pub async fn fetch_timetable_by_synergetic_id(synergetic_id: i32, user_id: i64, 
         // json[3][2]["ClassCodeDescription"]
         // json[5][2]["ClassCodeDescription"]
         let mut p_number = 1;
-        for j in 2..=9 { // No. 2,3,5,6,8,9 for periods 1,2,3,4,5,6 (4 & 7 are Lunch and Recess)
+        for j in 2..=9 {
+            // No. 2,3,5,6,8,9 for periods 1,2,3,4,5,6 (4 & 7 are Lunch and Recess)
             if j != 4 && j != 7 {
                 let class_name = json[j][i]["ClassCodeDescription"].to_string();
                 let teacher = json[j][i]["StaffName"].to_string();
@@ -99,13 +119,18 @@ pub async fn fetch_timetable_by_synergetic_id(synergetic_id: i32, user_id: i64, 
                 let period_number = p_number;
 
                 // Remove quotes
-                let class_name = &class_name[1..(class_name.len()-1)].to_string();
-                let teacher = &teacher[1..(teacher.len()-1)].to_string();
+                let class_name = &class_name[1..(class_name.len() - 1)].to_string();
+                let teacher = &teacher[1..(teacher.len() - 1)].to_string();
                 //let now = Instant::now();
                 class::create_class(
-                    timetable_id, day_number, period_number,
-                    class_name.clone(), teacher.clone(), pool
-                ).await?;
+                    timetable_id,
+                    day_number,
+                    period_number,
+                    class_name.clone(),
+                    teacher.clone(),
+                    pool,
+                )
+                .await?;
                 //println!("Pushed class [{}, {}] to database: {} μs", period_number, day_number, now.elapsed().as_micros());
 
                 p_number += 1;
@@ -114,24 +139,33 @@ pub async fn fetch_timetable_by_synergetic_id(synergetic_id: i32, user_id: i64, 
         d_number += 1;
     }
 
-    println!("Pushed classes to database for user {}: {}ms", user_id, now.elapsed().as_millis());
+    println!(
+        "Pushed classes to database for user {}: {}ms",
+        user_id,
+        now.elapsed().as_millis()
+    );
 
     Ok("successful".to_string())
 }
-
-
 
 //////////////////
 // CRUD Functions
 /////////////////
 
 // Create
-pub async fn create_timetable(user_id: i64, fetched_date: NaiveDate, pool: &Pool<Postgres>) -> Result<i32> {
+pub async fn create_timetable(
+    user_id: i64,
+    fetched_date: NaiveDate,
+    pool: &Pool<Postgres>,
+) -> Result<i32> {
     let timetable_id = sqlx::query!(
         "INSERT INTO timetables (user_id, fetched_date)
          VALUES ($1, $2) RETURNING id",
-        user_id, fetched_date
-    ).fetch_one(pool).await?;
+        user_id,
+        fetched_date
+    )
+    .fetch_one(pool)
+    .await?;
 
     Ok(timetable_id.id)
 }
@@ -142,17 +176,26 @@ pub async fn get_timetable_by_user_id(user_id: i64, pool: &Pool<Postgres>) -> Re
         Timetable,
         "SELECT * FROM timetables WHERE user_id = $1",
         user_id
-    ).fetch_one(pool).await?;
+    )
+    .fetch_one(pool)
+    .await?;
 
     Ok(timetable)
 }
 
 // Update
-pub async fn update_timetable_by_user_id(user_id: i64, fetched_date: NaiveDate, pool: &Pool<Postgres>) -> Result<()> {
+pub async fn update_timetable_by_user_id(
+    user_id: i64,
+    fetched_date: NaiveDate,
+    pool: &Pool<Postgres>,
+) -> Result<()> {
     sqlx::query!(
         "UPDATE timetables SET fetched_date = $1 WHERE user_id = $2",
-        fetched_date, user_id
-    ).execute(pool).await?;
+        fetched_date,
+        user_id
+    )
+    .execute(pool)
+    .await?;
 
     Ok(())
 }
@@ -160,27 +203,28 @@ pub async fn update_timetable_by_user_id(user_id: i64, fetched_date: NaiveDate, 
 // Delete - Also delete all classed associated with the timetable
 pub async fn delete_timetable_by_user_id(user_id: i64, pool: &Pool<Postgres>) -> Result<()> {
     // Get timetable id to use in deleting classes
-    let timetable_id = sqlx::query!(
-        r#"SELECT id FROM timetables WHERE user_id = $1"#,
-        user_id
-    ).fetch_one(pool).await?.id;
+    let timetable_id = sqlx::query!(r#"SELECT id FROM timetables WHERE user_id = $1"#, user_id)
+        .fetch_one(pool)
+        .await?
+        .id;
 
     class::delete_all_classes_in_timetable(timetable_id, &pool).await?;
 
-    sqlx::query!(
-        "DELETE FROM timetables WHERE user_id = $1",
-        user_id
-    ).execute(pool).await?;
+    sqlx::query!("DELETE FROM timetables WHERE user_id = $1", user_id)
+        .execute(pool)
+        .await?;
 
     Ok(())
 }
 
 // Check if timetable exists and return its id. If it doesn't exists, the returned id is 0
-pub async fn get_timetable_id_by_user_id_if_it_exists(user_id: i64, pool: &Pool<Postgres>) -> Result<i32> {
-    let records = sqlx::query!(
-        "SELECT id FROM timetables WHERE user_id = $1",
-        user_id
-    ).fetch_all(pool).await?; // Should only be returning one record
+pub async fn get_timetable_id_by_user_id_if_it_exists(
+    user_id: i64,
+    pool: &Pool<Postgres>,
+) -> Result<i32> {
+    let records = sqlx::query!("SELECT id FROM timetables WHERE user_id = $1", user_id)
+        .fetch_all(pool)
+        .await?; // Should only be returning one record
 
     if records.len() == 0 {
         return Ok(0);
